# openapi.yaml
openapi: 3.0.0
info:
  title: Mitosis API
  description: |
    Mitosis is a unified distributed transport evaluation framework that provides REST API endpoints 
    for managing users, groups, workers, and tasks in a distributed computing environment.

    ## Authentication
    Most endpoints require JWT bearer token authentication. Get a token by calling `/login` with 
    valid credentials.

    ## API Structure
    - `/auth` - User authentication status
    - `/login` - User login
    - `/users` - User management
    - `/groups` - Group management and file attachments
    - `/workers` - Worker registration and management
    - `/tasks` - Task submission and monitoring
    - `/admin` - Administrative operations (admin users only)
    - `/health` - Service health check
    - `/redis` - Redis connection information
  version: 0.6.0
  contact:
    name: Mitosis
    url: https://github.com/stack-rs/mitosis
  license:
    name: Apache 2.0
    url: https://github.com/stack-rs/mitosis/blob/main/LICENSE
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: https://api.mitosis.example.com
    description: Production server
paths:
...
  /workers:
    post:
      summary: Register worker
      description: Register a new worker with the coordinator
      operationId: registerWorker
      tags:
        - Workers
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterWorkerReq"
      responses:
        "200":
          description: Worker registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterWorkerResp"
        "401":
          $ref: "#/components/responses/Unauthorized"

    delete:
      summary: Unregister worker
      description: Unregister worker (worker auth required)
      operationId: unregisterWorker
      tags:
        - Workers
      security:
        - workerAuth: []
      responses:
        "200":
          description: Worker unregistered successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
  /workers/query:
    post:
      summary: Query workers
      description: List workers with optional filtering
      operationId: queryWorkers
      tags:
        - Workers
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkersQueryReq"
      responses:
        "200":
          description: List of workers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkersQueryResp"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /workers/shutdown:
    post:
      summary: Batch shutdown workers
      description: Shutdown multiple workers by filter criteria
      operationId: shutdownWorkersByFilter
      tags:
        - Workers
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkersShutdownByFilterReq"
      responses:
        "200":
          description: Workers shutdown successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkersShutdownByFilterResp"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /workers/shutdown/list:
    post:
      summary: Batch shutdown workers by UUIDs
      description: Shutdown multiple workers by specific UUIDs
      operationId: shutdownWorkersByUuids
      tags:
        - Workers
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkersShutdownByUuidsReq"
      responses:
        "200":
          description: Workers shutdown successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkersShutdownByUuidsResp"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /workers/{uuid}:
    get:
      summary: Get worker information
      description: Get detailed information about a specific worker
      operationId: getWorker
      tags:
        - Workers
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Worker information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerQueryResp"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      summary: Shutdown worker
      description: Request worker shutdown
      operationId: shutdownWorker
      tags:
        - Workers
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: op
          in: query
          required: false
          schema:
            type: string
            enum:
              - graceful
              - force
          description: |
            Worker shutdown operation:
            - Graceful: Wait for current task to complete
            - Force: Stop immediately
      responses:
        "200":
          description: Shutdown request sent
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /workers/{uuid}/tags:
    put:
      summary: Update worker tags
      description: Replace worker's tags
      operationId: updateWorkerTags
      tags:
        - Workers
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplaceWorkerTagsReq"
      responses:
        "200":
          description: Tags updated successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /workers/{uuid}/labels:
    put:
      summary: Update worker labels
      description: Replace worker's labels
      operationId: updateWorkerLabels
      tags:
        - Workers
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplaceWorkerLabelsReq"
      responses:
        "200":
          description: Tags updated successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /workers/{uuid}/groups:
    put:
      summary: Update worker group roles
      description: Add groups to worker or update their roles
      operationId: updateWorkerGroups
      tags:
        - Workers
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGroupWorkerRoleReq"
      responses:
        "200":
          description: Group roles updated successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      summary: Remove worker group roles (query params)
      description: Remove groups from worker using query parameters
      operationId: removeWorkerGroupsParams
      tags:
        - Workers
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: groups
          in: query
          required: false
          style: form
          explode: false
          schema:
            type: array
            items:
              type: string
          description: Comma-separated list of group names to remove
      responses:
        "200":
          description: Groups removed successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # /workers/{uuid}/groups/remove:
  #   post:
  #     summary: Remove worker group roles (JSON body)
  #     description: Remove groups from worker using JSON request body
  #     operationId: removeWorkerGroups
  #     tags:
  #       - Workers
  #     security:
  #       - bearerAuth: []
  #     parameters:
  #       - name: uuid
  #         in: path
  #         required: true
  #         schema:
  #           type: string
  #           format: uuid
  #     requestBody:
  #       required: true
  #       content:
  #         application/json:
  #           schema:
  #             $ref: "#/components/schemas/RemoveGroupWorkerRoleReq"
  #     responses:
  #       "200":
  #         description: Groups removed successfully
  #       "401":
  #         $ref: "#/components/responses/Unauthorized"
  #       "403":
  #         $ref: "#/components/responses/Forbidden"
  #       "404":
  #         $ref: "#/components/responses/NotFound"
  /workers/heartbeat:
    post:
      summary: Worker heartbeat
      description: Worker sends heartbeat to coordinator
      operationId: workerHeartbeat
      tags:
        - Workers
      security:
        - workerAuth: []
      responses:
        "200":
          description: Heartbeat received
        "401":
          $ref: "#/components/responses/Unauthorized"
  /workers/tasks:
    get:
      summary: Fetch task for worker
      description: Worker fetches next available task
      operationId: fetchTask
      tags:
        - Workers
      security:
        - workerAuth: []
      responses:
        "200":
          description: Task assigned to worker
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerTaskResp"
        "204":
          description: No tasks available
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      summary: Report task status
      description: Worker reports task execution status
      operationId: reportTask
      tags:
        - Workers
      security:
        - workerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportTaskReq"
      responses:
        "200":
          description: Status reported successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportTaskResp"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /workers/tasks/{uuid}:
    get:
      summary: Get task details for worker
      description: Worker queries task details
      operationId: getTaskForWorker
      tags:
        - Workers
      security:
        - workerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerTaskResp"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /workers/tasks/{uuid}/artifacts/{content_type}:
    get:
      summary: Download task artifact for worker
      description: Worker downloads task artifacts
      operationId: downloadTaskArtifactForWorker
      tags:
        - Workers
      security:
        - workerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: content_type
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/ArtifactContentType"
      responses:
        "302":
          description: Redirect to download URL
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /workers/tasks/{uuid}/attachments/{key}:
    get:
      summary: Download attachment for worker
      description: Worker downloads group attachments needed for task
      operationId: downloadAttachmentForWorker
      tags:
        - Workers
      security:
        - workerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: key
          in: path
          required: true
          schema:
            type: string
          style: simple
          explode: false
      responses:
        "302":
          description: Redirect to download URL
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /tasks:
    post:
      summary: Submit task
      description: Submit a new task for execution
      operationId: submitTask
      tags:
        - Tasks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitTaskReq"
      responses:
        "200":
          description: Task submitted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitTaskResp"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /tasks/query:
    post:
      summary: Query tasks
      description: List tasks with optional filtering
      operationId: queryTasks
      tags:
        - Tasks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TasksQueryReq"
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TasksQueryResp"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /tasks/{uuid}:
    get:
      summary: Get task information
      description: Get detailed information about a specific task
      operationId: getTask
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Task information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskQueryResp"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      summary: Update task
      description: Modify task configuration (only for pending tasks)
      operationId: updateTask
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeTaskReq"
      responses:
        "200":
          description: Task updated successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

    delete:
      summary: Cancel task
      description: Cancel a pending or running task
      operationId: cancelTask
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Task cancelled successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /tasks/{uuid}/labels:
    put:
      summary: Update task labels
      description: Replace task labels
      operationId: updateTaskLabels
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskLabelsReq"
      responses:
        "200":
          description: Labels updated successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /tasks/{uuid}/artifacts/{content_type}:
    delete:
      summary: Delete task artifact
      description: Delete a task artifact
      operationId: deleteTaskArtifact
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: content_type
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/ArtifactContentType"
      responses:
        "200":
          description: Artifact deleted successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /admin/workers/{uuid}/:
    delete:
      summary: Shutdown worker (Admin)
      description: Force shutdown a worker (admin only)
      operationId: adminShutdownWorker
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: op
          in: query
          required: false
          schema:
            type: string
            enum:
              - graceful
              - force
          description: |
            Worker shutdown operation:
            - Graceful: Wait for current task to complete
            - Force: Stop immediately
      responses:
        "200":
          description: Worker shutdown initiated
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /login endpoint
    workerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained when registering worker
  responses:
...
    Unauthorized:
      description: Unauthorized - invalid or missing authentication
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Invalid or expired token
...
  schemas:
...
    RegisterWorkerReq:
      type: object
      required:
        - tags
        - groups
      properties:
        tags:
          type: array
          items:
            type: string
          description: Worker capability tags
          example:
            - gpu
            - cuda
            - python
        groups:
          type: array
          items:
            type: string
          description: Groups this worker can serve
          example:
            - research-team
            - ml-group
        lifetime:
          type: string
          nullable: true
          description: Worker lifetime (e.g. "7d", "1h")
          example: 7d
    RegisterWorkerResp:
      type: object
      required:
        - worker_id
        - token
      properties:
        worker_id:
          type: string
          format: uuid
          description: Unique worker identifier
          example: 123e4567-e89b-12d3-a456-426614174000
        token:
          type: string
          description: JWT token for worker authentication
        redis_url:
          type: string
          nullable: true
          description: Redis URL for worker notifications
          example: redis://:password@localhost:6379
...
security:
  - bearerAuth: []
...